requirements.txt
alembic==1.13.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.25
pydantic==2.6.4
mysqlclient==2.2.4   # or use async driver: aiomysql
alembic==1.13.1      # for migrations
pytest==8.3.2        # testing
python-dotenv==1.0.1 
pymysql
jinja2
python-dotenv
python-multipart
playwright
pytest-cov
httpx